style: clippy + rustfmt cleanup across benches/tests/examples/src
Mechanical lint pass (cargo clippy --fix + cargo fmt) over the whole
crate. No behavioral changes:

- Remove redundant #![cfg(feature)] inner attrs (parent mod decls in
  lib.rs already gate salience / screening / alien_sampler).
- Drop unnecessary borrows and redundant 'as u32' casts.
- Add #[allow(clippy::too_many_arguments)] on numerical kernels whose
  arg count is intrinsic to their API surface.
- manual_range_contains, or_fun_call, needless borrows, if_same_then_else.
- rustfmt re-wrapping (derive macros, doc comments, signatures).
- Drop dead locals (ucb1_count) and fix overflowing literal test seed.

Verified: cargo check --all-features passes (only pre-existing warnings).

+++ src/salience/pending.rs
@@ -138,7 +138,7 @@ impl<A: Clone, const CAP: usize> PendingDelegateQueue<A, CAP> {
             return None;
         }
         // Oldest live slot = (head + CAP - len) mod CAP.
-        let oldest = ((self.head as usize + CAP - self.len as usize) % CAP) as usize;
+        let oldest = (self.head as usize + CAP - self.len as usize) % CAP;
         self.len -= 1;
         self.slots[oldest].take()
     }
